# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Система управления задачами - API
  description: |
    Микросервисная архитектура для управления пользователями и заказами
    в строительных проектах. ООО "СистемаКонтроля".
  version: 1.0.0
  contact:
    name: ООО СистемаКонтроля
    email: support@sistemakontrolya.ru
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: http://test-server:8000/api/v1
    description: Test server
  - url: https://api.production.com/api/v1
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен, полученный при входе
  schemas:
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Некорректные данные"
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "Иван Иванов"
        roles:
          type: array
          items:
            type: string
            enum: [user, admin, manager]
          example: ["user"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10
...
paths:
  /users/register:
    post:
      summary: Регистрация нового пользователя
      description: Создает нового пользователя в системе. Не требует авторизации.
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                  format: email
                  example: "newuser@example.com"
                password:
                  type: string
                  minLength: 6
                  example: "securePass123"
                name:
                  type: string
                  minLength: 2
                  example: "Петр Петров"
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Ошибка валидации или пользователь уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/login:
    post:
      summary: Вход пользователя
      description: Аутентификация пользователя и получение JWT токена
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  example: "password123"
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      user:
                        $ref: '#/components/schemas/User'
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/profile:
    get:
      summary: Получение профиля текущего пользователя
      description: Возвращает информацию о текущем авторизованном пользователе
      tags: [Users]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      summary: Обновление профиля
      description: Обновляет имя и/или email текущего пользователя
      tags: [Users]
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 2
                  example: "Новое Имя"
                email:
                  type: string
                  format: email
                  example: "newemail@example.com"
      responses:
        '200':
          description: Профиль обновлен
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users:
    get:
      summary: Список пользователей (только для администраторов)
      description: Возвращает список всех пользователей с пагинацией и поиском
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Номер страницы
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Количество записей на странице
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: search
          in: query
          description: Поиск по email или имени
          schema:
            type: string
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      users:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
        '403':
          description: Недостаточно прав (требуется роль admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...